requirements.txt
dify_plugin
python-docx
dataclasses